requirements.txt
python-chess==0.29.0
pycairo==1.18.2